instructorchat/requirements.txt
certifi==2025.6.15
colpali_engine==0.3.11
deepeval==3.2.6
deprecated==1.2.18
django_haystack==3.3.0
fast_plaid==1.0.3
openai==1.95.1
pandas==2.3.1
pdf2image==1.17.0
Pillow==11.3.0
pydantic==2.11.7
pymongo==4.13.2
pymupdf4llm==0.0.26
python-dotenv==1.1.1
rerankers==0.10.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
trio==0.30.0
trio_websocket==0.12.2
websockets==15.0.1
